test: fix paths
