+++ pkg/oci/pool.go
@@ -1,4 +1,4 @@
-package overlay
+package oci
 
 import (
 	"context"
